requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
pydantic
pydantic-settings